+++ src/dawn/native/ComputePassEncoder.cpp
@@ -468,7 +468,8 @@ void ComputePassEncoder::RestoreCommandBufferState(CommandBufferStateTracker sta
             if (offsets.empty()) {
                 APISetBindGroup(static_cast<uint32_t>(i), bg);
             } else {
-                APISetBindGroup(static_cast<uint32_t>(i), bg, offsets.size(), offsets.data());
+                APISetBindGroup(static_cast<uint32_t>(i), bg,
+                                checked_cast<uint32_t>(offsets.size()), offsets.data());
             }
         }
     }
