+++ src/dawn/native/TintUtils.cpp
@@ -165,7 +165,7 @@ tint::transform::VertexPulling::Config BuildVertexPullingTransformConfig(
         tint::transform::VertexBufferLayoutDescriptor* tintInfo =
             &cfg.vertex_state[static_cast<uint8_t>(slot)];
 
-        tintInfo->array_stride = dawnInfo.arrayStride;
+        tintInfo->array_stride = checked_cast<uint32_t>(dawnInfo.arrayStride);
         tintInfo->step_mode = ToTintVertexStepMode(dawnInfo.stepMode);
     }
 
@@ -174,7 +174,7 @@ tint::transform::VertexPulling::Config BuildVertexPullingTransformConfig(
         const VertexAttributeInfo& dawnInfo = renderPipeline.GetAttribute(location);
         tint::transform::VertexAttributeDescriptor tintInfo;
         tintInfo.format = ToTintVertexFormat(dawnInfo.format);
-        tintInfo.offset = dawnInfo.offset;
+        tintInfo.offset = checked_cast<uint32_t>(dawnInfo.offset);
         tintInfo.shader_location = static_cast<uint32_t>(static_cast<uint8_t>(location));
 
         uint8_t vertexBufferSlot = static_cast<uint8_t>(dawnInfo.vertexBufferSlot);
